college rep panel see extended description.
Sections "upload" and "p1" are the only ones live. Dashboard section will contain all the tiles (for events). I'll be adding the basic functionalities for the different sections meanwhile you guys look into the front-end code add something if needed. PHP code is not our concern I added those just to link up with the local database to see the front-end code working (it will be changed later so don't bother), link it up with your own local database to see it working.